+++ tests/Parsing/PartialParserTest.php
@@ -543,6 +543,32 @@ SOURCE;
     /**
      * @return void
      */
+    public function testGetLastNodeAtCorrectlyDealsWithHeredocContainingInterpolatedValues()
+    {
+        $source = <<<'SOURCE'
+<?php
+
+<<<EOF
+EOF: {$foo[0]} some_text
+
+This is / some text.
+
+EOF
+SOURCE;
+
+        $result = $this->createPartialParser()->getLastNodeAt($source);
+
+        $this->assertInstanceOf(Node\Scalar\Encapsed::class, $result);
+        $this->assertInstanceOf(Node\Scalar\EncapsedStringPart::class, $result->parts[0]);
+        $this->assertEquals('EOF: ', $result->parts[0]->value);
+        $this->assertInstanceOf(Node\Expr\ArrayDimFetch::class, $result->parts[1]);
+        $this->assertInstanceOf(Node\Scalar\EncapsedStringPart::class, $result->parts[2]);
+        $this->assertEquals(" some_text\n\nThis is / some text.\n", $result->parts[2]->value);
+    }
+
+    /**
+     * @return void
+     */
     public function testGetLastNodeAtCorrectlyDealsWithSpecialClassConstantClassKeyword()
     {
         $source = <<<'SOURCE'