+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 2.1.3
+* Fix corner case with HEREDOCs containing interpolated values failing to parse, causing indexing to fail for files containg them (https://github.com/php-integrator/core/issues/82).
 * Default value parsing failures will now throw `LogicException`s.
   * This will cause them to crash the server, but that way they can be debugged as parsing valid PHP code should never fail.