+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## 3.0.0
 * At least PHP 7.1 is now required.
 
+## 2.1.3
+* Fix corner case with HEREDOCs containing interpolated values failing to parse, causing indexing to fail for files containg them (https://github.com/php-integrator/core/issues/82).
+* Default value parsing failures will now throw `LogicException`s.
+  * This will cause them to crash the server, but that way they can be debugged as parsing valid PHP code should never fail.
+
 ## 2.1.2
 * Fix `@throws` tags without a description being ignored.
 * Fix symlinks not being followed in projects that have them.